+++ examples/context_menu/main.go
@@ -28,14 +28,11 @@ func main() {
 }
 
 func mainView(w *gui.Window) gui.View {
-	ww, wh := w.WindowSize()
 	app := gui.State[App](w)
 
 	return gui.ContextMenu(w, gui.ContextMenuCfg{
 		ID:     "ctx",
-		Width:  float32(ww),
-		Height: float32(wh),
-		Sizing: gui.FixedFixed,
+		Sizing: gui.FillFill,
 		HAlign: gui.HAlignCenter,
 		VAlign: gui.VAlignMiddle,
 		Items: []gui.MenuItemCfg{