+++ examples/native_menu/main.go
@@ -102,14 +102,11 @@ func registerCommands(w *gui.Window) {
 }
 
 func mainView(w *gui.Window) gui.View {
-	ww, wh := w.WindowSize()
 	app := gui.State[App](w)
 	theme := gui.CurrentTheme()
 
 	return gui.Column(gui.ContainerCfg{
-		Width:  float32(ww),
-		Height: float32(wh),
-		Sizing: gui.FixedFixed,
+		Sizing: gui.FillFill,
 		HAlign: gui.HAlignCenter,
 		Content: []gui.View{
 			gui.Rectangle(gui.RectangleCfg{